Update README.md
I am  going to divide the whole webpage into two sections. The left side has a menu called Documentation Menu with all the topics listed inside a navigation bar. It contains anchor tags that have links to the IDs of all the sections(like encapsulation, polymorphism, etc.). On the right side, we have descriptions for each of the topics. The idea behind this is once the user clicks on one of the topics in the left section, respective topic details will load the content on the right. In the CSS file, we are just beautifying the text like its alignment, padding margin, etc.